+++ halgo/Halgo/CLA/Command/Transaction.hs
@@ -13,15 +13,15 @@ import Control.Monad (forM_, when)
 import Options.Applicative (Parser, command, hsubparser, info, progDesc)
 import UnliftIO (liftIO)
 
-import qualified Data.Algorand.Address as A
-import qualified Data.Algorand.Amount as A
-import qualified Data.Algorand.Transaction as T
-import qualified Data.Algorand.Transaction.Build as T
-import qualified Data.Algorand.Transaction.Group as T
-import qualified Data.Algorand.Transaction.Signed as TS
 import qualified Network.Algorand.Api as Api
 
+import Data.Algorand.Address (Address, zero)
 import Data.Algorand.Amount (Microalgos)
+import Data.Algorand.Asset (AssetIndex)
+import Data.Algorand.Transaction (Transaction, TransactionType (..), transactionId)
+import Data.Algorand.Transaction.Build (buildTransaction)
+import Data.Algorand.Transaction.Group (isValidGroup, makeGroup)
+import Data.Algorand.Transaction.Signed (signFromContractAccount, signSimple, verifyTransaction)
 import Network.Algorand.Definitions (Host)
 
 import Halgo.CLA.Argument (argAddress, argAmount, argAssetIndex, argProgramFile, argSecretFile)
@@ -76,7 +76,7 @@ cmdTxnShow :: MonadSubCommand m => Bool -> Bool -> Bool -> m ()
 cmdTxnShow verify json base64 = do
   stxns <- if json then readItemsJson else readItemsB64
   when verify $
-    forM_ stxns $ \stxn -> case TS.verifyTransaction stxn of
+    forM_ stxns $ \stxn -> case verifyTransaction stxn of
       Nothing -> die "Invalid signature. Run with --no-verify if you still want to see it."
       Just _txn -> pure ()
   (if base64 then putItemsB64 else putItemsJson) stxns
@@ -84,7 +84,7 @@ cmdTxnShow verify json base64 = do
 -- | Show unsigned transactions.
 cmdTxnShowUnsigned :: MonadSubCommand m => Bool -> Bool -> m ()
 cmdTxnShowUnsigned json base64 =
-  (if json then readItemsJson else readItemsB64 @T.Transaction)
+  (if json then readItemsJson else readItemsB64 @Transaction)
   >>= if base64 then putItemsB64 else putItemsJson
 
 -- | Sign transactions with a simple signature.
@@ -92,7 +92,7 @@ cmdTxnSign :: MonadSubCommand m => Bool -> FilePath -> m ()
 cmdTxnSign json skFile = do
   sk <- loadAccount skFile
   txns <- if json then readItemsJson else readItemsB64
-  let signed = map (TS.signSimple sk) txns
+  let signed = map (signSimple sk) txns
   putItemsB64 signed
 
 -- | Sign transactions with a logic signature.
@@ -100,35 +100,37 @@ cmdTxnLogicSign :: MonadSubCommand m => Bool -> FilePath -> m ()
 cmdTxnLogicSign json programFile = do
   program <- liftIO $ BS.readFile programFile
   txns <- if json then readItemsJson else readItemsB64
-  let signed = map (TS.signFromContractAccount program []) txns
+  let signed = map (signFromContractAccount program []) txns
   putItemsB64 signed
 
 -- | Transaction ID.
 cmdTxnId :: MonadSubCommand m => Bool -> m ()
 cmdTxnId json = do
   txns <- if json then readItemsJson else readItemsB64
-  mapM_ (putTextLn . T.transactionId) txns
+  mapM_ (putTextLn . transactionId) txns
 
 -- | Create or check a group of transactions
 cmdTxnGroup :: MonadSubCommand m => Bool -> Bool -> m ()
 cmdTxnGroup json check = do
   txns <- if json then readItemsJson else readItemsB64
   case check of
-    True -> case T.isValidGroup txns of
+    True -> case isValidGroup txns of
       False -> die "Not a valid transaction group"
       True -> putItemsB64 txns
-    False -> putItemsB64 $ T.makeGroup txns
+    False -> putItemsB64 $ makeGroup txns
 
-cmdTxnNewPay :: MonadSubCommand m => A.Address -> A.Microalgos -> Host -> m ()
+cmdTxnNewPay :: MonadSubCommand m => Address -> Microalgos -> Host -> m ()
 cmdTxnNewPay to amnt url = withNode url $ \(_, api) -> do
   params <- Api._transactionsParams api
-  let payment = T.PaymentTransaction to amnt Nothing
-  let txn = T.buildTransaction params A.zero payment
+  let payment = PaymentTransaction to amnt Nothing
+  let txn = buildTransaction params zero payment
   putItemsB64 [txn]
 
-cmdTxnNewAxfr :: MonadSubCommand m => T.AssetIndex -> A.Address -> Microalgos -> Host -> m ()
+cmdTxnNewAxfr
+  :: MonadSubCommand m
+  => AssetIndex -> Address -> Microalgos -> Host -> m ()
 cmdTxnNewAxfr index to amnt url = withNode url $ \(_, api) -> do
   params <- Api._transactionsParams api
-  let payment = T.AssetTransferTransaction index amnt Nothing to Nothing
-  let txn = T.buildTransaction params A.zero payment
+  let payment = AssetTransferTransaction index amnt Nothing to Nothing
+  let txn = buildTransaction params zero payment
   putItemsB64 [txn]